delete EmitModuleSwappingConfigPass
EmitModuleSwappingConfig functionality folded into CreateSiFiveMetadata
